EncryptedBackupViewModel refactored
